updated code
Done:

User Login and SignUp (with Basic Details) (Mandatory)

You can log in via the email field, You need to manage sessions. (Mandatory)

Manage Blogs, CRUD. (Mandatory)
Ability to change the status of the post (Active/Inactive) and accordingly it will reflect inside the post listing.  (Mandatory)

Use any editor for adding and editing blogs (Preferable - editor.js ).  (Mandatory)

Show a list of blogs as an infinite loading on the homepage with author name and number of comments on that. (Mandatory)